feat(plugin): 添加插件配置注解和配置管理功能
- 添加 PluginConfig 注解用于插件配置定义
- 在 PluginDescribe 注解中添加 enable 和 enableWeb 属性
- 修改数据库实体类将 ID 类型从 Integer 改为 Long
- 新增 CreateConfigRequest DTO 类用于创建配置请求
- 实现配置注册、查询、更新等核心功能
- 添加插件数据表和相关实体类及服务类
- 更新 SQL 映射文件和数据库 schema 定义
- 添加配置键值校验逻辑确保格式正确
- 在心跳处理器中添加 echo 消息日志记录

+++ MangoBot-cli/src/main/java/io/github/mangomaner/mangobot/controller/BotConfigController.java
@@ -65,7 +65,7 @@ public class BotConfigController {
     @GetMapping("/id/{id}")
     @Operation(summary = "根据配置ID获取配置")
     public BaseResponse<ConfigVO> getConfigById(
-            @Parameter(description = "配置ID") @PathVariable Integer id) {
+            @Parameter(description = "配置ID") @PathVariable Long id) {
         ConfigVO config = mangobotConfigService.getConfigById(id);
         return ResultUtils.success(config);
     }
@@ -78,7 +78,7 @@ public class BotConfigController {
     @PostMapping("/AiModel/test")
     @Operation(summary = "测试Ai模型")
     public BaseResponse<AiTestResultVo> testAiModel(
-            @Parameter(description = "配置ID") @RequestParam Integer id) {
+            @Parameter(description = "配置ID") @RequestParam Long id) {
         String result = mangobotConfigService.getConfigById(id).getConfigValue();
         AiTestResultVo aiTestResultVo = new AiTestResultVo();
         aiTestResultVo.setSuccess(true);
